+++ examples/webhook-signing/README.md
@@ -58,6 +58,10 @@ To add a new example:
 3. Create a `main.ts` file with sample server code and run `chmod +x main.ts` so it's directly executable.
 4. The `main.ts` has to follow some rules:
   1. Written in typescript using `import .. from ..`
-  1. Must have `#!/usr/bin/env npx ts-node` header to be directly runable.
-  1. Must print `Webhook endpoint available at <url>` when the server is started.
-  1. Must return `{"received":true}` and 200 status code for `charge.succeeded` event.
+  2. Must have `#!/usr/bin/env npx ts-node` header to be directly runable.
+  3. Must print `Webhook endpoint available at <url>` when the server is started.
+  4. Must return `{"received":true}` and 200 status code for `charge.succeeded` event.
+5. Test you example
+  1. `cd examples/webhook-signing/test`
+  2. `./main.ts ../<your test directory>`
+  3. Add a test to `/tests/Integration.spec.ts`
