+++ test/stripe.spec.ts
@@ -659,58 +659,4 @@ describe('Stripe Module', function() {
       expect(newStripe.VERSION).to.equal(Stripe.PACKAGE_VERSION);
     });
   });
-
-  describe('imports', function() {
-    this.timeout(30000);
-    const runTestProject = (projectName: string): void => {
-      const script = `
-        cd testProjects/${projectName} &&
-        npm install &&
-        npm run lint &&
-        npm run runtestproject -- ${testUtils.getUserStripeKey()}
-      `;
-      require('child_process').execSync(script);
-    };
-
-    it('should work with CommonJS imports', () => {
-      expect(runTestProject.bind(null, 'cjs')).to.not.throw();
-    });
-
-    it('should work with ESModule imports', function() {
-      // Node supports ES Modules starting at v12
-      if (parseInt(process.versions.node.split('.')[0], 10) <= 12) {
-        this.skip();
-      }
-
-      expect(runTestProject.bind(null, 'mjs')).to.not.throw();
-      expect(runTestProject.bind(null, 'mjs-ts')).to.not.throw();
-    });
-  });
-
-  describe('stripe.worker', function() {
-    if (process.versions.node < '16.13') {
-      console.log('Wrangler requires at least node.js v16.13.0, skipping test');
-      return;
-    }
-    const runTestCloudflareProject = (projectName: string): void => {
-      const script = `
-        cd testProjects/${projectName} &&
-        npm install &&
-        npm run build
-      `;
-      require('child_process').execSync(script);
-    };
-
-    it('should build successfully in Cloudflare Workers', function() {
-      expect(
-        runTestCloudflareProject.bind(null, 'cloudflare-worker')
-      ).to.not.throw();
-    });
-
-    it('should build successfully in Cloudflare Pages Functions', function() {
-      expect(
-        runTestCloudflareProject.bind(null, 'cloudflare-pages')
-      ).to.not.throw();
-    });
-  });
 });
